requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
alembic>=1.13.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.1.0
apscheduler>=3.10.4
python-dateutil>=2.9.0
httpx>=0.27.0
cryptography>=42.0.0
python-keycloak>=4.5.0
